# testdata/lso/pkg/lso/server.gen.go
// Code generated by MarkRosemaker/openapi-codegen. DO NOT EDIT.

package lso

import (
	"context"
	"encoding/json/v2"
	"errors"
	"fmt"
	"log/slog"
	"net/http"
	"strconv"

	"github.com/go-api-libs/api/server"
)

// Service defines the operations the server must implement.
type Service interface {
	ListAccounts(ctx context.Context) (*ListAccountsResult, error)
	GetPositions(ctx context.Context, accountID string) (*Positions, error)
	GetCash(ctx context.Context, accountID string) (*GetCashResult, error)
	GetAccountSummary(ctx context.Context, accountID string) (*AccountSummary, error)
	GetFundamentals(ctx context.Context, cik int) (*FundamentalsAnalysis, error)
	GetCompanyBySymbol(ctx context.Context, symbol string) (*CompanyAndSecurity, error)
	GetIncomeStatement(ctx context.Context, cik int) (*IncomeStatementAnalysis, error)
	GetBalanceSheet(ctx context.Context, cik int) (*BalanceSheetAnalysis, error)
	GetWatchlist(ctx context.Context) (*Watchlist, error)
	GetMarketStatus(ctx context.Context) (*MarketStatus, error)
}

// RegisterService registers a [Service] with an [*http.ServeMux].
// Optionally, a path prefix may be provided.
func RegisterService(svc Service, mux *http.ServeMux, pathPrefix string) {
	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/accounts")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "ListAccounts"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			res, err := svc.ListAccounts(ctx)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}

	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/accounts/{account_id}/positions")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "GetPositions"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			accountID := r.PathValue("account_id")

			res, err := svc.GetPositions(ctx, accountID)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}

	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/accounts/{account_id}/cash")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "GetCash"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			accountID := r.PathValue("account_id")

			res, err := svc.GetCash(ctx, accountID)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}

	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/accounts/{account_id}/summary")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "GetAccountSummary"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			accountID := r.PathValue("account_id")

			res, err := svc.GetAccountSummary(ctx, accountID)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}

	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/fundamentals/{cik}")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "GetFundamentals"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			s := r.PathValue("cik")
			cik, err := strconv.Atoi(s)
			if err != nil {
				msg := fmt.Sprintf("invalid cik: %v", err)
				l.DebugContext(ctx, "Bad Request", slog.String("msg", msg))
				http.Error(w, msg, http.StatusBadRequest)
				return
			}

			res, err := svc.GetFundamentals(ctx, cik)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}

	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/fundamentals/{symbol}/company")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "GetCompanyBySymbol"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			symbol := r.PathValue("symbol")

			res, err := svc.GetCompanyBySymbol(ctx, symbol)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}

	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/fundamentals/{cik}/income-statement")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "GetIncomeStatement"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			s := r.PathValue("cik")
			cik, err := strconv.Atoi(s)
			if err != nil {
				msg := fmt.Sprintf("invalid cik: %v", err)
				l.DebugContext(ctx, "Bad Request", slog.String("msg", msg))
				http.Error(w, msg, http.StatusBadRequest)
				return
			}

			res, err := svc.GetIncomeStatement(ctx, cik)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}

	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/fundamentals/{cik}/balance-sheet")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "GetBalanceSheet"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			s := r.PathValue("cik")
			cik, err := strconv.Atoi(s)
			if err != nil {
				msg := fmt.Sprintf("invalid cik: %v", err)
				l.DebugContext(ctx, "Bad Request", slog.String("msg", msg))
				http.Error(w, msg, http.StatusBadRequest)
				return
			}

			res, err := svc.GetBalanceSheet(ctx, cik)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}

	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/watchlist")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "GetWatchlist"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			res, err := svc.GetWatchlist(ctx)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}

	{
		path := fmt.Sprintf("%s%s", pathPrefix, "/market/status")
		l := slog.Default().With(slog.String("method", "GET"), slog.String("path", path), slog.String("function", "GetMarketStatus"))

		mux.HandleFunc(fmt.Sprintf("GET %s", path), func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			l.DebugContext(ctx, "called")

			res, err := svc.GetMarketStatus(ctx)
			if err != nil {
				sErr, ok := errors.AsType[*server.Error](err)
				if !ok {
					l.ErrorContext(ctx, "Internal Server Error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
					return
				}

				l.DebugContext(ctx, "graceful error", slog.String("error", err.Error()))

				w.Header().Set("Content-Type", "application/json")
				w.WriteHeader(sErr.Code)

				if err := json.MarshalWrite(w, err); err != nil {
					l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				}
				return
			}

			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusOK)

			if err := json.MarshalWrite(w, res, jsonOpts); err != nil {
				l.ErrorContext(ctx, "marshal error", slog.String("error", err.Error()))
				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
				return
			}

			l.DebugContext(ctx, "success")
		})
	}
}